move versioned manifest to manifest/mod.rs
